Move contract calls into separate module
